Otimização na criação de moedas e espinhos
